api/requirements.txt
fastapi
pyjwt
tortoise-orm
uvicorn[standard]
passlib
bcrypt
pyjwt
python-multipart
python-dotenv
pytest
pydantic==1.10.11
mkdocs
pandas
aerich
polars==0.16.11
pika
tortoise-orm[asyncpg]
mkdocstrings[python]
mkdocs-material
mkdocs-static-i18nt
pyarrow